+++ eth_defi/erc_4626/vault_protocol/spectra/__init__.py
@@ -1,0 +1,1 @@
+"""Spectra Finance vault protocol support."""